+++ client/src/components/UI/Header.module.scss
@@ -16,3 +16,15 @@
     opacity: 0.9;
   }
 }
+
+.menu {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  a {
+    color: white;
+    font-size: 1.2rem;
+    padding-right: 1rem;
+  }
+}
