+++ client/src/pages/Auth.js
@@ -1,5 +1,4 @@
 import React, { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
 import { useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import * as yup from 'yup';
@@ -27,7 +26,6 @@ let cx = classNames.bind(classes);
 const Auth = () => {
   const [isLoginMode, setIsLoginMode] = useState(true);
   const dispatch = useDispatch();
-  const navigate = useNavigate();
 
   const signupClass = cx({
     signup: true,
